pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clipjits"
version = "0.1.0"
description = "A two-step CLI tool for creating BJJ technique cards from instructional videos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp>=2024.0.0",
    "ffmpeg-python>=0.2.0",
    "openai-whisper>=20231117",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
clipjits = "clipjits.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["clipjits*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
